+++ CHANGELOG.md
@@ -5,9 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.5.9-24] - 2020-07-27
 ### Added
-- StoryConfig: taginfos
+- StoryConfig: taginfos, chars and papers
 - Outputter: datetime print on last
+- Runner: output chars on console
 ### Changed
 - Counter: is_contain_material
 - Writer: into world.py
@@ -616,7 +619,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Example story as usage.
 - Output story as markdown.
 
-[Unreleased]: https://github.com/nagisc007/storybuilder/compare/v0.5.9-23...HEAD
+[Unreleased]: https://github.com/nagisc007/storybuilder/compare/v0.5.9-24...HEAD
+[0.5.9-24]: https://github.com/nagisc007/storybuilder/releases/v0.5.9-24
 [0.5.9-23]: https://github.com/nagisc007/storybuilder/releases/v0.5.9-23
 [0.5.9-22]: https://github.com/nagisc007/storybuilder/releases/v0.5.9-22
 [0.5.9-21]: https://github.com/nagisc007/storybuilder/releases/v0.5.9-21